refactor(finance-imports): decompose correction proposal + tag review
