+++ packages/app-finance/src/components/imports/CorrectionProposalDialog.tsx
@@ -1,41 +1,10 @@
-import { Plus, Search } from 'lucide-react';
-import { lazy, Suspense, useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { toast } from 'sonner';
-
-import { trpc } from '@pops/api-client';
-import { Button, Input, WorkflowDialog } from '@pops/ui';
-
-import {
-  applyBrowsePriorityReorder,
-  sortRulesForBrowseDisplay,
-} from '../../lib/correction-browse-reorder';
-import { computeMergedRules } from '../../lib/merged-state';
-import { useImportStore } from '../../store/importStore';
 import {
   type CorrectionSignal,
-  type LocalOp,
   type ServerChangeSet,
   type TriggeringTransactionContext,
 } from './correction-proposal-shared';
-import {
-  AiHelperPanel,
-  BrowseRuleDetailPanel,
-  ContextPanel,
-  DetailPanel,
-  ImpactPanel,
-  OpsListPanel,
-  type PreviewView,
-  RejectPanel,
-} from './CorrectionProposalDialogPanels';
-import { useApplyRejectMutations } from './hooks/useApplyRejectMutations';
-import { localOpsToChangeSet, newClientId, useLocalOps } from './hooks/useLocalOps';
-import { usePreviewEffects } from './hooks/usePreviewEffects';
-
-import type { CorrectionRule } from './RulePicker';
-
-const BrowseRulesSidebar = lazy(() =>
-  import('./BrowseRulesSidebar').then((m) => ({ default: m.BrowseRulesSidebar }))
-);
+import { CorrectionProposalWorkflow } from './correction-proposal/CorrectionProposalWorkflow';
+import { CorrectionRuleManagerDialog } from './correction-proposal/CorrectionRuleManagerDialog';
 
 // Re-export shared symbols so existing consumers don't break
 export type {
@@ -78,620 +47,29 @@ export interface CorrectionProposalDialogProps {
   onBrowseClose?: (hadChanges: boolean) => void;
 }
 
-function previewLabel(view: PreviewView, hasSelectedOp: boolean): string {
-  if (view === 'combined') return 'Combined effect of entire ChangeSet';
-  if (hasSelectedOp) return 'Effect of selected operation';
-  return 'No operation selected';
-}
-
-// ---------------------------------------------------------------------------
-// Main component
-// ---------------------------------------------------------------------------
-
 export function CorrectionProposalDialog(props: CorrectionProposalDialogProps) {
   const minConfidence = props.minConfidence ?? 0.7;
-  const isBrowseMode = props.mode === 'browse';
-  const { onOpenChange, onBrowseClose } = props;
-
-  // ---- initial propose query ---------------------------------------------
-
-  const disabledSignal: CorrectionSignal = useMemo(
-    () => ({ descriptionPattern: '_', matchType: 'exact', tags: [] }),
-    []
-  );
-
-  const proposeInput = useMemo(() => {
-    if (!props.signal) return null;
-    return { signal: props.signal, minConfidence, maxPreviewItems: 200 };
-  }, [props.signal, minConfidence]);
-
-  const proposeQuery = trpc.core.corrections.proposeChangeSet.useQuery(
-    proposeInput ?? { signal: disabledSignal, minConfidence, maxPreviewItems: 200 },
-    {
-      enabled: Boolean(!isBrowseMode && props.open && proposeInput),
-      staleTime: 0,
-      retry: false,
-    }
-  );
-
-  // ---- extracted hooks ---------------------------------------------------
-
-  const localOpsHook = useLocalOps({
-    open: props.open,
-    signal: props.signal,
-    isBrowseMode,
-    proposeData: proposeQuery.data,
-  });
-
-  const {
-    localOps,
-    setLocalOps,
-    selectedClientId,
-    setSelectedClientId,
-    selectedOp,
-    rationale,
-    setRationale,
-    updateOp,
-    handleDeleteOp,
-    handleAddNewRuleOp,
-    handleAddTargetedOp,
-    seededForSignalRef,
-  } = localOpsHook;
-
-  const pendingChangeSets = useImportStore((s) => s.pendingChangeSets);
-
-  // ---- browse mode: fetch DB transactions for impact preview (PRD-032 US-06) ---
-
-  const dbTxnsQuery = trpc.finance.transactions.listDescriptionsForPreview.useQuery(undefined, {
-    enabled: isBrowseMode && props.open,
-    staleTime: 60_000,
-  });
-
-  const previewHook = usePreviewEffects(
-    {
-      open: props.open,
-      localOps,
-      selectedOp,
-      minConfidence,
-      previewTransactions: props.previewTransactions,
-      dbTransactions: isBrowseMode ? (dbTxnsQuery.data?.data ?? []) : undefined,
-      pendingChangeSets,
-    },
-    setLocalOps
-  );
-
-  const {
-    combinedPreview,
-    combinedPreviewError,
-    combinedPreviewTruncated,
-    combinedDbPreview,
-    selectedOpPreview,
-    selectedOpPreviewError,
-    selectedOpPreviewTruncated,
-    selectedOpDbPreview,
-    previewMutationPending,
-    hasDirty,
-    handleRerunPreview,
-    resetPreviewState,
-    lastCombinedStructuralSigRef,
-    selectedOpPreviewKeyRef,
-  } = previewHook;
-
-  const handleCloseRef = useRef<() => void>(() => {});
-
-  const mutationsHook = useApplyRejectMutations({
-    signal: props.signal,
-    sessionId: props.sessionId,
-    localOps,
-    combinedPreview,
-    combinedPreviewError,
-    previewTransactions: props.previewTransactions,
-    isFetching: proposeQuery.isFetching,
-    previewMutationPending,
-    hasDirty,
-    onApproved: props.onApproved,
-    onClose: () => {
-      handleCloseRef.current();
-    },
-    setLocalOps,
-    setSelectedClientId,
-    setRationale,
-    lastCombinedStructuralSigRef,
-    selectedOpPreviewKeyRef,
-  });
-
-  const {
-    rejectMode,
-    setRejectMode,
-    rejectFeedback,
-    setRejectFeedback,
-    aiInstruction,
-    setAiInstruction,
-    aiMessages,
-    aiBusy,
-    isBusy,
-    canApply,
-    handleApprove,
-    handleConfirmReject,
-    handleAiSubmit,
-    resetMutationState,
-  } = mutationsHook;
-
-  // ---- preview view state ------------------------------------------------
-  const [previewView, setPreviewView] = useState<PreviewView>('selected');
-
-  // ---- browse mode state --------------------------------------------------
-  const [browseSearch, setBrowseSearch] = useState('');
-  const [browseSelectedRuleId, setBrowseSelectedRuleId] = useState<string | null>(null);
-  const browseInitialPendingCountRef = useRef<number>(0);
-  const addPendingChangeSet = useImportStore((s) => s.addPendingChangeSet);
-
-  const browseListQuery = trpc.core.corrections.list.useQuery(
-    { limit: 500, offset: 0 },
-    { enabled: isBrowseMode && props.open, staleTime: 30_000 }
-  );
-
-  const browseMergedRules: CorrectionRule[] = useMemo(() => {
-    if (!isBrowseMode) return [];
-    const browseDbRules = browseListQuery.data?.data ?? [];
-    if (pendingChangeSets.length === 0) return browseDbRules;
-    return computeMergedRules(browseDbRules, pendingChangeSets);
-  }, [isBrowseMode, browseListQuery.data?.data, pendingChangeSets]);
-
-  const browseOrderedMerged = useMemo(
-    () => sortRulesForBrowseDisplay(browseMergedRules, localOps),
-    [browseMergedRules, localOps]
-  );
-
-  const browseOrderedFiltered = useMemo(() => {
-    const needle = browseSearch.trim().toLowerCase();
-    if (!needle) return browseOrderedMerged;
-    return browseOrderedMerged.filter((r) => {
-      const haystack =
-        `${r.descriptionPattern} ${r.entityName ?? ''} ${r.matchType} ${r.location ?? ''}`.toLowerCase();
-      return haystack.includes(needle);
-    });
-  }, [browseOrderedMerged, browseSearch]);
-
-  const browseCanDragReorder = browseSearch.trim() === '' && browseOrderedMerged.length >= 2;
-
-  const handleBrowseReorderFullList = useCallback(
-    (reordered: CorrectionRule[]) => {
-      setLocalOps((prev) => applyBrowsePriorityReorder(reordered, prev));
-    },
-    [setLocalOps]
-  );
-
-  const browseSelectedRule = useMemo(
-    () => browseMergedRules.find((r) => r.id === browseSelectedRuleId) ?? null,
-    [browseMergedRules, browseSelectedRuleId]
-  );
-
-  useEffect(() => {
-    if (isBrowseMode && props.open) {
-      browseInitialPendingCountRef.current = useImportStore.getState().pendingChangeSets.length;
-    }
-  }, [isBrowseMode, props.open]);
-
-  // ---- handlers -----------------------------------------------------------
-
-  const handleOpenChange = useCallback(
-    (open: boolean) => {
-      if (!open && isBrowseMode) {
-        const currentCount = useImportStore.getState().pendingChangeSets.length;
-        const hadChanges = currentCount !== browseInitialPendingCountRef.current;
-        setBrowseSearch('');
-        setBrowseSelectedRuleId(null);
-        setLocalOps([]);
-        setSelectedClientId(null);
-        onOpenChange(false);
-        onBrowseClose?.(hadChanges);
-        return;
-      }
-      onOpenChange(open);
-      if (!open) {
-        setLocalOps([]);
-        setSelectedClientId(null);
-        setPreviewView('selected');
-        resetPreviewState();
-        resetMutationState();
-        setRationale(null);
-        seededForSignalRef.current = null;
-      }
-    },
-    [
-      isBrowseMode,
-      onOpenChange,
-      onBrowseClose,
-      resetPreviewState,
-      resetMutationState,
-      setLocalOps,
-      setSelectedClientId,
-      setRationale,
-      seededForSignalRef,
-    ]
-  );
-
-  handleCloseRef.current = () => {
-    handleOpenChange(false);
-  };
-
-  const handleBrowseSelectRule = useCallback(
-    (ruleId: string) => {
-      setBrowseSelectedRuleId(ruleId);
-      const existingOp = localOps.find((o) => o.kind !== 'add' && o.targetRuleId === ruleId);
-      if (existingOp) {
-        setSelectedClientId(existingOp.clientId);
-      } else {
-        setSelectedClientId(null);
-      }
-    },
-    [localOps, setSelectedClientId]
-  );
-
-  const handleBrowseEditRule = useCallback(
-    (rule: CorrectionRule) => {
-      const newOp: LocalOp = {
-        kind: 'edit',
-        clientId: newClientId('edit'),
-        targetRuleId: rule.id,
-        targetRule: rule,
-        data: {
-          entityId: rule.entityId ?? undefined,
-          entityName: rule.entityName ?? undefined,
-          location: rule.location ?? undefined,
-          tags: rule.tags,
-          transactionType: rule.transactionType ?? undefined,
-          isActive: rule.isActive,
-          confidence: rule.confidence,
-        },
-        dirty: true,
-      };
-      setLocalOps((prev) => [...prev, newOp]);
-      setSelectedClientId(newOp.clientId);
-    },
-    [setLocalOps, setSelectedClientId]
-  );
-
-  const handleBrowseDisableRule = useCallback(
-    (rule: CorrectionRule) => {
-      const newOp: LocalOp = {
-        kind: 'disable',
-        clientId: newClientId('disable'),
-        targetRuleId: rule.id,
-        targetRule: rule,
-        rationale: '',
-        dirty: true,
-      };
-      setLocalOps((prev) => [...prev, newOp]);
-      setSelectedClientId(newOp.clientId);
-    },
-    [setLocalOps, setSelectedClientId]
-  );
-
-  const handleBrowseRemoveRule = useCallback(
-    (rule: CorrectionRule) => {
-      const newOp: LocalOp = {
-        kind: 'remove',
-        clientId: newClientId('remove'),
-        targetRuleId: rule.id,
-        targetRule: rule,
-        rationale: '',
-        dirty: true,
-      };
-      setLocalOps((prev) => [...prev, newOp]);
-      setSelectedClientId(newOp.clientId);
-    },
-    [setLocalOps, setSelectedClientId]
-  );
-
-  const handleBrowseSave = useCallback(() => {
-    if (localOps.length === 0) {
-      handleOpenChange(false);
-      return;
-    }
-    const changeSet = localOpsToChangeSet(localOps, { source: 'browse-rule-manager' });
-    if (changeSet) {
-      addPendingChangeSet({ changeSet, source: 'browse-rule-manager' });
-      toast.success(`${localOps.length} rule change${localOps.length === 1 ? '' : 's'} saved`);
-    }
-    handleOpenChange(false);
-  }, [localOps, addPendingChangeSet, handleOpenChange]);
-
-  // ---- shared memos (must be above any early return) ----------------------
-
-  const excludeIds = useMemo(() => {
-    const set = new Set<string>();
-    for (const op of localOps) {
-      if (op.kind !== 'add') set.add(op.targetRuleId);
-    }
-    return set;
-  }, [localOps]);
-
-  // ---- render -------------------------------------------------------------
-
-  if (isBrowseMode) {
-    const browseFooter = (
-      <>
-        <div className="flex-1 text-xs text-muted-foreground">
-          {localOps.length > 0 && (
-            <span>
-              {localOps.length} unsaved change{localOps.length === 1 ? '' : 's'}
-            </span>
-          )}
-        </div>
-        <Button variant="outline" onClick={() => handleOpenChange(false)}>
-          Cancel
-        </Button>
-        <Button onClick={handleBrowseSave} disabled={localOps.length === 0}>
-          Save Changes
-        </Button>
-      </>
-    );
-
-    const browseBody = browseListQuery.isError ? (
-      <div className="px-6 pb-6 text-sm text-destructive">{browseListQuery.error.message}</div>
-    ) : browseListQuery.isLoading ? (
-      <div className="px-6 pb-6 text-sm text-muted-foreground">Loading rules…</div>
-    ) : (
-      // 3-column grid rendered via WorkflowDialog columns prop
-      <>
-        {/* Sidebar: rule list with search */}
-        <div className="flex flex-col min-h-0 border-r">
-          <div className="px-3 py-2 border-b">
-            <div className="relative">
-              <Search className="absolute left-2 top-2.5 h-3.5 w-3.5 text-muted-foreground" />
-              <Input
-                value={browseSearch}
-                onChange={(e) => {
-                  setBrowseSearch(e.target.value);
-                }}
-                placeholder="Search rules…"
-                className="pl-7 h-8 text-xs"
-              />
-            </div>
-          </div>
-          <div className="px-3 py-1.5 text-[10px] text-muted-foreground border-b space-y-0.5">
-            <div>
-              {browseOrderedFiltered.length} rule
-              {browseOrderedFiltered.length === 1 ? '' : 's'}
-              {browseSearch && ` matching "${browseSearch}"`}
-            </div>
-            {browseSearch.trim() !== '' && (
-              <div className="text-[10px] text-muted-foreground/90">
-                Clear search to drag rules into priority order.
-              </div>
-            )}
-          </div>
-          <div className="flex-1 overflow-auto">
-            <Suspense fallback={<div className="p-3 text-xs text-muted-foreground">Loading…</div>}>
-              <BrowseRulesSidebar
-                canDragReorder={browseCanDragReorder}
-                orderedMerged={browseOrderedMerged}
-                orderedFiltered={browseOrderedFiltered}
-                selectedRuleId={browseSelectedRuleId}
-                localOps={localOps}
-                onSelectRule={handleBrowseSelectRule}
-                onReorderFullList={handleBrowseReorderFullList}
-              />
-            </Suspense>
-          </div>
-          <div className="border-t p-2">
-            <Button
-              size="sm"
-              variant="outline"
-              className="w-full justify-start"
-              onClick={handleAddNewRuleOp}
-            >
-              <Plus className="mr-1 h-3.5 w-3.5" /> Add new rule
-            </Button>
-          </div>
-        </div>
-
-        {/* Detail: show selected rule or selected op editor */}
-        <div className="flex flex-col min-h-0 overflow-auto">
-          {selectedOp ? (
-            <DetailPanel
-              op={selectedOp}
-              onChange={(mutator) => {
-                if (!selectedOp) return;
-                updateOp(selectedOp.clientId, mutator);
-              }}
-              disabled={false}
-            />
-          ) : browseSelectedRule ? (
-            <BrowseRuleDetailPanel
-              rule={browseSelectedRule}
-              onEdit={handleBrowseEditRule}
-              onDisable={handleBrowseDisableRule}
-              onRemove={handleBrowseRemoveRule}
-            />
-          ) : (
-            <div className="p-6 text-sm text-muted-foreground">
-              Select a rule on the left to view or edit it.
-            </div>
-          )}
-        </div>
-
-        {/* Impact preview (PRD-032 US-06) */}
-        {(() => {
-          const browsePreviewResult =
-            previewView === 'combined' ? combinedPreview : selectedOpPreview;
-          const browseDbPreviewResult =
-            previewView === 'combined' ? combinedDbPreview : selectedOpDbPreview;
-          const browsePreviewError =
-            previewView === 'combined' ? combinedPreviewError : selectedOpPreviewError;
-          const browseTruncated =
-            previewView === 'combined' ? combinedPreviewTruncated : selectedOpPreviewTruncated;
-          const browseLabel =
-            previewView === 'combined'
-              ? 'Combined effect of all pending changes'
-              : selectedOp
-                ? 'Effect of selected operation'
-                : 'No operation selected';
-          return (
-            <ImpactPanel
-              view={previewView}
-              onViewChange={setPreviewView}
-              label={browseLabel}
-              previewResult={browsePreviewResult}
-              dbPreviewResult={browseDbPreviewResult}
-              dbTruncated={dbTxnsQuery.data?.truncated}
-              dbTotal={dbTxnsQuery.data?.total}
-              previewError={browsePreviewError}
-              isPending={previewMutationPending}
-              stale={hasDirty}
-              truncated={browseTruncated}
-              onRerun={handleRerunPreview}
-              disabled={localOps.length === 0}
-            />
-          );
-        })()}
-      </>
-    );
-
-    const isGridMode = !browseListQuery.isError && !browseListQuery.isLoading;
-
+  if (props.mode === 'browse') {
     return (
-      <WorkflowDialog
+      <CorrectionRuleManagerDialog
         open={props.open}
-        onOpenChange={handleOpenChange}
-        title="Manage Rules"
-        description="Browse, search, and edit classification rules. Changes are buffered locally until import is committed."
-        columns={isGridMode ? 3 : undefined}
-        gridTemplate={isGridMode ? 'grid-cols-[300px_minmax(0,1fr)_360px]' : undefined}
-        footer={browseFooter}
-      >
-        {browseBody}
-      </WorkflowDialog>
+        onOpenChange={props.onOpenChange}
+        onBrowseClose={props.onBrowseClose}
+        minConfidence={minConfidence}
+      />
     );
   }
 
-  // ---- proposal mode render -----------------------------------------------
-
-  const previewResult = previewView === 'combined' ? combinedPreview : selectedOpPreview;
-  const previewError = previewView === 'combined' ? combinedPreviewError : selectedOpPreviewError;
-  const previewTruncated =
-    previewView === 'combined' ? combinedPreviewTruncated : selectedOpPreviewTruncated;
-  const currentPreviewLabel = previewLabel(previewView, !!selectedOp);
-
-  const proposalFooter = (
-    <>
-      <div className="flex-1 text-xs text-muted-foreground">
-        {hasDirty ? (
-          <span>Preview stale — re-run before applying.</span>
-        ) : localOps.length === 0 ? (
-          <span>ChangeSet is empty.</span>
-        ) : null}
-      </div>
-      <Button variant="outline" onClick={() => handleOpenChange(false)} disabled={isBusy}>
-        Cancel
-      </Button>
-      {!rejectMode && (
-        <Button
-          variant="outline"
-          onClick={() => setRejectMode(true)}
-          disabled={isBusy || localOps.length === 0}
-        >
-          Reject with feedback
-        </Button>
-      )}
-      <Button onClick={handleApprove} disabled={!canApply}>
-        Apply ChangeSet
-      </Button>
-    </>
-  );
-
-  const isProposalGridReady =
-    props.signal && !proposeQuery.isError && !(proposeQuery.isLoading && localOps.length === 0);
-
-  const proposalBody = !props.signal ? (
-    <div className="px-6 pb-6 text-sm text-muted-foreground">No proposal signal provided.</div>
-  ) : proposeQuery.isError ? (
-    <div className="px-6 pb-6 text-sm text-destructive">{proposeQuery.error.message}</div>
-  ) : proposeQuery.isLoading && localOps.length === 0 ? (
-    <div className="px-6 pb-6 text-sm text-muted-foreground">Generating proposal…</div>
-  ) : (
-    <>
-      <OpsListPanel
-        ops={localOps}
-        selectedClientId={selectedClientId}
-        onSelect={setSelectedClientId}
-        onDelete={handleDeleteOp}
-        onAddNewRule={handleAddNewRuleOp}
-        onAddTargeted={handleAddTargetedOp}
-        excludeIds={excludeIds}
-        disabled={isBusy}
-      />
-      <DetailPanel
-        op={selectedOp}
-        onChange={(mutator) => {
-          if (!selectedOp) return;
-          updateOp(selectedOp.clientId, mutator);
-        }}
-        disabled={isBusy}
-      />
-      <ImpactPanel
-        view={previewView}
-        onViewChange={setPreviewView}
-        label={currentPreviewLabel}
-        previewResult={previewResult}
-        previewError={previewError}
-        isPending={previewMutationPending}
-        stale={hasDirty}
-        truncated={previewTruncated}
-        onRerun={handleRerunPreview}
-        disabled={isBusy || localOps.length === 0}
-      />
-    </>
-  );
-
-  const proposalSubpanel =
-    isProposalGridReady &&
-    (rejectMode ? (
-      <RejectPanel
-        feedback={rejectFeedback}
-        onFeedbackChange={setRejectFeedback}
-        onCancel={() => {
-          setRejectMode(false);
-          setRejectFeedback('');
-        }}
-        onConfirm={handleConfirmReject}
-        busy={mutationsHook.rejectMutationPending}
-      />
-    ) : (
-      <AiHelperPanel
-        messages={aiMessages}
-        instruction={aiInstruction}
-        onInstructionChange={setAiInstruction}
-        onSubmit={handleAiSubmit}
-        busy={aiBusy}
-      />
-    ));
-
-  const proposalContextHeader =
-    isProposalGridReady && props.signal ? (
-      <ContextPanel
-        signal={props.signal}
-        triggeringTransaction={props.triggeringTransaction}
-        rationale={rationale}
-        opCount={localOps.length}
-        combinedSummary={combinedPreview?.summary ?? null}
-      />
-    ) : undefined;
-
   return (
-    <WorkflowDialog
+    <CorrectionProposalWorkflow
       open={props.open}
-      onOpenChange={handleOpenChange}
-      title="Correction proposal"
-      description="Edit the proposed rule changes and preview their impact before applying."
-      columns={isProposalGridReady ? 3 : undefined}
-      header={proposalContextHeader || undefined}
-      subpanel={proposalSubpanel || undefined}
-      footer={proposalFooter}
-    >
-      {proposalBody}
-    </WorkflowDialog>
+      onOpenChange={props.onOpenChange}
+      sessionId={props.sessionId}
+      signal={props.signal}
+      triggeringTransaction={props.triggeringTransaction}
+      previewTransactions={props.previewTransactions}
+      minConfidence={minConfidence}
+      onApproved={props.onApproved}
+    />
   );
 }
